mv params,mockOrder to test_common.go for testing
