+++ x/order/types/params.go
@@ -18,7 +18,6 @@ const (
 	DefaultFeeAmountPerBlock     = "0" // okt
 	DefaultFeeDenomPerBlock      = common.NativeToken
 	DefaultFeeRateTrade          = "0.001"    // percentage
-	DefaultTestFeeAmountPerBlock = "0.000001" // okt
 )
 
 // nolint : Parameter keys
@@ -28,7 +27,6 @@ var (
 	KeyFeePerBlock         = []byte("FeePerBlock")
 	KeyTradeFeeRate        = []byte("TradeFeeRate")
 	DefaultFeePerBlock     = sdk.NewDecCoinFromDec(DefaultFeeDenomPerBlock, sdk.MustNewDecFromStr(DefaultFeeAmountPerBlock))
-	DefaultTestFeePerBlock = sdk.NewDecCoinFromDec(DefaultFeeDenomPerBlock, sdk.MustNewDecFromStr(DefaultTestFeeAmountPerBlock))
 )
 
 // nolint
@@ -65,15 +63,6 @@ func DefaultParams() Params {
 		OrderExpireBlocks: DefaultOrderExpireBlocks,
 		MaxDealsPerBlock:  DefaultMaxDealsPerBlock,
 		FeePerBlock:       DefaultFeePerBlock,
-		TradeFeeRate:      sdk.MustNewDecFromStr(DefaultFeeRateTrade),
-	}
-}
-
-func DefaultTestParams() Params {
-	return Params{
-		OrderExpireBlocks: DefaultOrderExpireBlocks,
-		MaxDealsPerBlock:  DefaultMaxDealsPerBlock,
-		FeePerBlock:       DefaultTestFeePerBlock,
 		TradeFeeRate:      sdk.MustNewDecFromStr(DefaultFeeRateTrade),
 	}
 }
